requirements_cuda118.txt
absl-py==2.1.0
anyio==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-lru==2.0.4
attrs==23.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==4.1.0
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==5.0.1
comm==0.2.1
contourpy==1.0.5
cycler==0.11.0
Cython==3.0.10
data==0.4
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
easydict==1.13
einops==0.8.0
exceptiongroup==1.2.0
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.15.4
fonttools==4.51.0
fsspec==2024.6.1
funcsigs==1.0.2
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.31.0
google-auth-oauthlib==0.4.6
grpcio==1.64.1
huggingface-hub==0.23.4
idna==3.7
importlib-metadata==7.0.1
importlib_resources==6.4.0
ipykernel==6.28.0
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.4
jpeg4py==0.1.4
json5==0.9.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.0
jupyter_server==2.14.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.11
jupyterlab-pygments==0.1.2
jupyterlab_server==2.25.1
jupyterlab-widgets==3.0.10
kiwisolver==1.4.4
latex==0.7.0
lmdb==1.5.1
loguru==0.7.2
Markdown==3.6
MarkupSafe==2.1.3
matplotlib==3.4.0
matplotlib-inline==0.1.6
mistune==2.0.4
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpld3==0.5.9
mpmath==1.3.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.1
notebook==6.4.12
notebook_shim==0.2.3
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.10.0.84
overrides==7.4.0
packaging==24.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.43
protobuf==5.27.2
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pycparser==2.21
Pygments==2.15.1
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.30.2
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
rsa==4.9
safetensors==0.4.3
scipy==1.10.1
Send2Trash==1.8.2
sentry-sdk==2.7.1
setproctitle==1.3.3
setuptools==59.5.0
setuptools-scm==8.1.0
shutilwhich==1.1.0
sip==6.7.12
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.2.0
sympy==1.12.1
tempdir==0.7.1
tensorboard==2.12.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
terminado==0.17.1
tikzplotlib==0.10.1
timm==1.0.7
tinycss2==1.2.1
tomli==2.0.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
triton==2.3.1
typing_extensions==4.11.0
tzdata==2024.1
unicodedata2==15.1.0
urllib3==2.2.2
visdom==0.2.4
wandb==0.17.4
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.10
zipp==3.17.0
